+++ postcss.config.js
@@ -1,21 +1,14 @@
-const postcssImport = require('postcss-import');
-const nesting = require('postcss-nesting');
-const customMedia = require('postcss-custom-media');
-const customProperties = require('postcss-custom-properties');
-const calc = require('postcss-calc');
-const colorFunction = require('postcss-color-function');
-const flexbugsFixes = require('postcss-flexbugs-fixes');
-const autoprefixer = require('autoprefixer');
-
 module.exports = {
-  plugins: [
-    postcssImport(),
-    nesting(),
-    customMedia(),
-    customProperties(),
-    calc(),
-    colorFunction(),
-    flexbugsFixes(),
-    autoprefixer(),
-  ],
+  plugins: {
+    'postcss-import': {},
+    'postcss-nesting': {},
+    'postcss-custom-media': {},
+    'postcss-custom-properties': {
+      preserve: false,
+    },
+    'postcss-calc': {},
+    'postcss-color-function': {},
+    'postcss-flexbugs-fixes': {},
+    autoprefixer: {},
+  },
 };